requirements.txt
psycopg2-binary==2.9.1
python-dateutil==2.8.1
retry==0.9.2

numpy==2.0.2
pandas==2.2.3